Update index.html
Added a new <div> element with the id "loveFacts" to your existing HTML structure. This is where love facts will be displayed.
Created a placeholder section for love facts within the "loveFacts" div, where the facts will be dynamically inserted.
Linked this HTML structure to the JavaScript code, allowing it to display love feedback and facts after calculating the love percentage.